Move TLSerdes to its own file
